pyproject.toml
[project]
name = "datasaki-backend"
version = "0.1.0"
description = "DataSaki Backend API"
authors = [
    {name = "Kapil Jain",email = "kapiljain1989@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.109.2,<0.110.0)",
    "uvicorn (>=0.27.1,<0.28.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "authlib (>=1.5.1,<2.0.0)",
    "sqlalchemy (>=2.0.27,<3.0.0)",
    "alembic (>=1.13.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-jose (>=3.3.0,<4.0.0)",
    "python-multipart (>=0.0.9,<0.0.10)",
    "httpx (>=0.28.1,<0.29.0)",
    "psycopg2-binary (>=2.9.9,<3.0.0)",
    "pydantic (>=2.6.1,<3.0.0)",
    "oauthlib (>=3.2.2,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "bcrypt==3.2.2",
    "passlib==1.7.4",
    "pdfplumber (>=0.11.6,<0.12.0)",
    "boto3 (>=1.37.34,<2.0.0)",
    "pandas (>=2.2.0,<3.0.0)",
    "clickhouse-driver (>=0.2.9,<0.3.0)",
    "langchain (>=0.1.4,<0.2.0)",
    "langchain-core (>=0.1.17,<0.2.0)",
    "langchain-community (>=0.0.10,<0.1.0)",
    "langchain-openai (>=0.0.5,<0.1.0)",
    "langchain-anthropic (>=0.0.1,<0.1.0)",
    "langchain-google-genai (>=0.0.5,<0.1.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "pydantic-settings (>=2.1.0,<3.0.0)",
    "numpy (>=1.26.3,<2.0.0)",
    "openai (>=1.12.0,<2.0.0)",
    "anthropic (>=0.18.1,<0.19.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run-dev = "scripts.run_dev:main"
